add before & after to add spacing around hline

+++ src/main/groovy/pdfdsl/support/Location.groovy
@@ -56,7 +56,7 @@ class Location {
       result
     } else {
       invoked = cache
-      result = valueClosure(rect, mapWrapper)
+      result = valueClosure(rect, mapWrapper) as float
     }
   }
